requirements.txt
###############
# FastAPI + Server
###############
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pydantic
pydantic-settings

###############
# Vector DB - Pinecone
###############
pinecone   # última versión estable

###############
# RAG - Embeddings & NLP
###############
sentence-transformers
scikit-learn         # requerido por sentence-transformers
numpy
torch                # CPU/GPU auto-compatible

###############
# PDF Processing (deep analysis)
###############
pymupdf   # fitz, para extraer imágenes y texto
python-docx          # para DOCX

###############
# LLM Integrations
###############
openai
transformers
requests

###############
# LangChain (mínimo, sin LangGraph)
###############
langchain
langchain-core
langchain-community

###############
# AWS tools (opcionales en tu caso)
###############
boto3

###############
# Tests
###############
pytest
pytest-asyncio
httpx


